+++ setup.py
@@ -11,7 +11,7 @@ setup(name='mrs',
                 'mrs.allocation',
                 'mrs.tests',
                 'mrs.messages',
-                'mrs.scheduling',
+                'mrs.timetable',
                 'mrs.dispatching',
                 'mrs.execution',
                 'mrs.utils'],
